Moved `host/syscall_condition.rs` to `host/syscall/condition.rs`
