+++ src/main/host/mod.rs
@@ -9,6 +9,5 @@ pub mod network;
 pub mod process;
 pub mod status_listener;
 pub mod syscall;
-pub mod syscall_condition;
 pub mod thread;
 pub mod timer;
